chore: review fixes — DRY auth helper extraction
Move authorizePublish() to shared registry/lib/auth.ts,
eliminating duplicated auth+permission pattern across
[...name].ts and index.ts.
